Refactor: RequestParser 클래스를 RequestHandler에서 분리
